requirements.txt
click
lxml
pip-chill
pyexecjs
requests
selenium
